requirements.txt
# DIPS-Plus requirements
setuptools==56.2.0
dill==0.3.3
tqdm==4.49.0
Sphinx==4.0.1
easy-parallel-py3==0.1.6.4
atom3-py3==0.1.8.7
click==7.0.0
# mpi4py==3.0.3  # On Andes, do 'source venv/bin/activate', 'module load gcc/10.3.0', and 'pip install mpi4py --no-cache-dir --no-binary :all:'
# pytorch-lightning==1.3.2  # Install only after PyTorch (w/ CUDA support) is installed
